Ids in special offers
